ai-worker/requirements.txt
boto3
pyannote.audio
yt-dlp
openai
openai-whisper
opencv-python
redis
python-dotenv
requests
numpy
